Moved some files to the loader/cache subdirectory
